requirements.txt
dvc==2.34.0
dvc==s3==2.23.0
fsspec==2024.9.0
lightning==2.3.3
mlflow==2.17.0
omegaconf==2.3.0
opencv==python==4.10.0.84
Pillow==10.4.0
python==dotenv==1.0.1
pytorch-ignite==0.5.1
torch==2.4.1
torchmetrics==1.4.3 
torchvision==0.19.1 
tqdm==4.66.5

